prac_8/ecg_analysis_service/requirements.txt
python3-pika
torch
torch-ecg
numpy
pandas
neurokit2
scipy
scikit-learn
pika  # RabbitMQ client
minio
python-dotenv
peakutils
openai
requests
